apis handle worked hours

+++ MinimalAPI/Data/WorkedHours.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public string Person { get; set; }
         public DateTime Day { get; set; }
-        public double HoursAmount { get; set; }
+        public float HoursAmount { get; set; }
     }
 }
